One Blob Implementation
This is the one blob encoding. It is a subclass of torch.nn.Modules, just import and use as a torch.nn module. You need to define the upper_bound and the lower_bound
This module takes in a batch of 3D coordinates(of size (B,3) ) within the bound, and then returns an encoded tensor of size (B, 3*num_bins).